move parseSkipMask to its own file, extract arr fn to utils
